--- src/main/java/ru/geekbrains/lim/task1/Obstruction.java
@@ -1,5 +1,0 @@
-package ru.geekbrains.lim.task1;
-
-public interface Obstruction {
-    boolean overcome(Participant participant);
-}
