+++ src/main/java/ru/geekbrains/lim/task1/obstruction/Obstruction.java
@@ -1,0 +1,7 @@
+package ru.geekbrains.lim.task1.obstruction;
+
+import ru.geekbrains.lim.task1.participant.Participant;
+
+public interface Obstruction {
+    boolean overcome(Participant participant);
+}
