+++ src/main/java/ru/geekbrains/lim/task1/participant/Participant.java
@@ -1,4 +1,4 @@
-package ru.geekbrains.lim.task1;
+package ru.geekbrains.lim.task1.participant;
 
 public interface Participant {
     boolean run(int distance);
