+++ src/main/java/ru/geekbrains/lim/task1/participant/Robot.java
@@ -1,24 +1,24 @@
-package ru.geekbrains.lim.task1;
+package ru.geekbrains.lim.task1.participant;
 
-public class Cat implements Participant {
+public class Robot implements Participant {
 
     private final int maxRunDistance;
     private final int maxJumpHeight;
 
-    public Cat(int maxRunDistance, int maxJumpHeight) {
+    public Robot(int maxRunDistance, int maxJumpHeight) {
         this.maxRunDistance = maxRunDistance;
         this.maxJumpHeight = maxJumpHeight;
     }
 
     @Override
     public boolean run(int distance) {
-        System.out.print("Кот пробежал ");
+        System.out.print("Робот пробежал ");
         return maxRunDistance >= distance;
     }
 
     @Override
     public boolean jump(int height) {
-        System.out.print("Кот прыгнул ");
+        System.out.print("Робот прыгнул ");
         return maxJumpHeight >= height;
     }
 }
